airflow/pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aws>=0.2.5",
    "boto3>=1.36.5",
    "loguru>=0.7.3",
    "pillow>=11.1.0",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "apache-airflow>=2.10.4",
    "aiohttp>=3.11.11",
]
